Removed prune phase and "on" member from quadtree
The prune phase was unnecessary, and could probably cause a segfault if
the render thread got hung while on a node that got deleted by the
search thread.  Now, nodes that would be pruned are simply never placed
in the tree and are deleted immediately after it is determined that they
should be pruned.

Also removed the node.on member - it is now implicitly encoded.  Nodes
that used to have node->on == NODE_OFF are deleted before being placed
in the tree.  Nodes that used to have node->on == NODE_CHILDON have
node->totalchildren > 0, and nodes that used to have node->on == NODE_ON
have node->totalchildren == 0.